Split of docker utils. Automagically pull docker image
