+++ PyStemmusScope/bmi/docker_process.py
@@ -1,71 +1,17 @@
 """The Docker STEMMUS_SCOPE model process wrapper."""
 import os
-import warnings
-from pathlib import Path
 from time import sleep
 from typing import Any
 from PyStemmusScope.config_io import read_config
+from PyStemmusScope.bmi.docker_utils import find_image, check_tags, make_docker_vols_binds
 
 
 try:
     import docker
+    import docker.errors
 except ImportError:
     docker = None
-
-
-def make_docker_vols_binds(cfg_file: str) -> tuple[list[str], list[str]]:
-    """Make docker volume mounting configs.
-
-    Args:
-        cfg_file: Location of the config file
-
-    Returns:
-        volumes, binds
-    """
-    cfg = read_config(cfg_file)
-    cfg_dir = Path(cfg_file).parent
-    volumes = []
-    binds = []
-
-    # Make sure no subpaths are mounted:
-    if not cfg_dir.is_relative_to(cfg["InputPath"]):
-        volumes.append(str(cfg_dir))
-        binds.append(f"{str(cfg_dir)}:{str(cfg_dir)}")
-    if (not Path(cfg["InputPath"]).is_relative_to(cfg_dir)) or (
-        Path(cfg["InputPath"]) == cfg_dir
-    ):
-        volumes.append(cfg["InputPath"])
-        binds.append(f"{cfg['InputPath']}:{cfg['InputPath']}")
-    if not Path(cfg["OutputPath"]).is_relative_to(cfg_dir):
-        volumes.append(cfg["OutputPath"])
-        binds.append(f"{cfg['OutputPath']}:{cfg['OutputPath']}")
-
-    return volumes, binds
-
-
-def check_tags(image: str, compatible_tags: tuple[str, ...]):
-    """Check if the tag is compatible with this version of the BMI.
-
-    Args:
-        image: The full image name (including tag)
-        compatible_tags: Tags which are known to be compatible with this version of the
-            BMI.
-    """
-    if ":" not in image:
-        msg = (
-            "Could not validate the Docker image tag, as no tag was provided.\n"
-            "Please set the Docker image tag in the configuration file."
-        )
-        warnings.warn(UserWarning(msg), stacklevel=1)
-
-    tag = image.split(":")[-1]
-    if tag not in compatible_tags:
-        msg = (
-            f"Docker image tag '{tag}' not found in compatible tags "
-            f"({compatible_tags}).\n"
-            "You might experience issues or unexpected results."
-        )
-        warnings.warn(UserWarning(msg), stacklevel=1)
+    docker.errors = None
 
 
 def wait_for_model(phrase: bytes, socket: Any) -> None:
@@ -96,6 +42,7 @@ class StemmusScopeDocker:
         config = read_config(cfg_file)
 
         self.image = config["DockerImage"]
+        find_image(self.image)
         check_tags(self.image, self.compatible_tags)
 
         self.client = docker.APIClient()
